Render NotFoundPage for a missing canvas
