+++ app/src/shared/errors/EditorError.js
@@ -1,0 +1,9 @@
+// @flow
+
+export default class EditorError {
+    source: any
+
+    constructor(source: any) {
+        this.source = source
+    }
+}
